fix: add missing playlist_length
Also convert to TS
